fix: make Potion.Service compile with 0 errors
Resolve 340+ compile errors so 'dotnet build src/Potion.Service' succeeds:
- add Program.cs entrypoint (Generic Host + WebHost + Windows Service)
- migrate Polly 8 APIs (OnTimeout/OnOpened/OnRejected, args.Timeout)
- drop chaos pipeline block (Polly.Simmy does not exist on nuget.org);
  implement bulkhead via Polly.RateLimiting ConcurrencyLimiter
- fix ML.NET 5.0 API drift (DetectIidSpike transform, RandomizedPca)
- deduplicate types (RemediationTaskDescriptor, UserSession)
- resolve CS860x nullable warnings and add missing members/properties
- add Polly.RateLimiting + Microsoft.Extensions.Hosting.WindowsServices
- add src/Potion.Service/README.md (referenced by csproj)
- untrack committed obj/ artifacts and extend .gitignore

+++ src/Potion.Service/Infrastructure/ChaosEngineeringService.cs
@@ -442,7 +442,7 @@ public class DiskStressFaultInjector : IFaultInjector
             while (DateTimeOffset.UtcNow < endTime && !cancellationToken.IsCancellationRequested)
             {
                 var fileName = Path.Combine(tempDir, $"chaos-test-{fileIndex}.tmp");
-                var data = new byte[1024 * intensity]; // サイズはintensityに基づく
+                var data = new byte[(int)(1024 * intensity)]; // サイズはintensityに基づく
 
                 for (int i = 0; i < data.Length; i++)
                 {
